requirements.txt
Django==5.2.8
mysqlclient==2.2.6
python-dotenv==1.0.0
